requirements.txt
fastmcp>=0.1.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pyjwt>=2.8.0
aiosqlite>=0.19.0